refactor overlays
